Add a lance benchmark
