+++ vortex-ipc/Cargo.toml
@@ -16,6 +16,7 @@ flatbuffers = { workspace = true }
 itertools = { workspace = true }
 vortex-array = { path = "../vortex-array" }
 vortex-error = { path = "../vortex-error", features = ["flatbuffers"] }
+vortex-flatbuffers = { path = "../vortex-flatbuffers" }
 vortex-schema = { path = "../vortex-schema" }
 
 [build-dependencies]
