--- vortex-ipc/src/stream/mod.rs
@@ -1,2 +1,0 @@
-pub mod reader;
-pub mod writer;
